requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.4
filelock==3.0.12
importlib-metadata==4.8.1
mysqlclient==2.0.3
Pillow==8.3.2
platformdirs==2.2.0
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
virtualenv==20.7.2
whitenoise==5.3.0
zipp==3.5.0
